move utility functions into utils module
